% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading the bz2 zipped files with corresponding year in their name and
selecting only months and years of this dataset.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list/vector containing the years for which data should be
loaded and month + year are extracted}
}
\value{
A dataframe containing the month and the year of each observation
        of the zipped file - the year coming as input and has to be part of
        the file name.
}
\description{
The function \code{fars_read_years} can load several files corresponding to
their year into R, since it gets a list/vector of years as input and returns
the month and years of that dataframes.
It is still possible to only load one dataset and manipulate it.
The lapply functionality allows to create for each element (year of years)
the zipped file name, check whether it exists with \code{tryCatch} and if it
throws an error, write out a warning that the year is invalid and return
\code{NULL}.
If the file exists, the inner of \code{tryCatch} is executed and the data is
loaded into R. Then a new column year is created, which is the current
element of years chosen in lapply and only Month and year of the dataset are
selected.
}
\details{
Errors occur in the same cases as for \code{make_filename}, but not for
\code{fars_read}, since this is part of the tryCatch.
}
\examples{
# fars_read_years(list(2013,2014,2015))
# fars_read_years(2013)
}
