% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Reading bz2 zipped data files and creating a summary of the number of
observations per month per year (years correspond to datasets)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list/vector containing the years for which data should be
loaded and month + year are extracted}
}
\value{
A dataframe containing the number of each month and the number of
        observations for each month in each of the input years
}
\description{
The function \code{fars_summarize_years} uses the \code{fars_read_years}
function toload the datasets and creates a list of each dataset containing
only the month and year per observation.
All this data is combined, via binding all rows together.
Next, the data is grouped by month and year that the number of observations
of each grouping can be measured.
In the last step, year and the number of observations per month and year
\code{(=n)} are spread, meaning new columns are created for each element of
the year column and it's value is the number of \code{n} in this month.
}
\details{
Error occur, in the same cases as for \code{fars_read_years}.
If there are no errors in this help function, everything works well.
}
\examples{
# fars_summarize_years(list(2013,2014,2015))
# fars_summarize_years(2013)
}
