% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS State Maps}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer corresponding to the number of the state
in the data set.}

\item{year}{Either string or integer representation of the year
for which the map is to be drawn.}
}
\value{
NULL; side effect: plots the map of the state.
}
\description{
Map of a state with plotting all accidents for the given year.
}
\details{
For each year and state number passed into the function, create a
    map of the state showing the location accidents in that year.
}
