% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plot the accident on the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a list of integers or character string of numbers for desired
states.}

\item{year}{a list of integer or character string of numbers for desired years
to be read from the disk.}
}
\value{
a map of accidents.
}
\description{
The function filters for state.num and year given in the input. if the number
of accident is zero for given year and state it returns "no accident to plot".
The function plot the map with accident as point on it.
}
\examples{

}

