% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading a csv file into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string containing the path for the file}
}
\value{
A tibble dataframe containing the information of the csv file
}
\description{
The function \code{fars_read} gets as input the path of the file name to load
into R. It checks whether the file exists and if not, stops.
If the file exists, with correct format, it is loaded into R with the
\code{read_csv} function from the \code{readr} package and messages,which
type each column is, are suppressed.
At least, the function uses the \code{tbl_df} function from the \code{dplyr}
package, although there is no need to do that.
}
\details{
Erros occur by files which exist but are not in (zipped) csv format, or files
which are in csv format but do not have the default delimiters.
Error occurs if more than one filename is passed.
}
\examples{
# filename <- "inst/extdata/accident_2013.csv.bz2"
# accident_2013 <- fars_read(filename)
}
