% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{FARS Read Funcion}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{String contianing the name and relative path to a
file.}
}
\value{
A "tbl_df" tibble object.
}
\description{
Read CSV files into a tibble.  Using \code{\link[dplyr]{tbl_df}}
read a file supported by that function.
}
