% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{reads csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string giving the filename.}
}
\value{
returns a dataframe if the file exist on the disk.
}
\description{
This function load csv file from the disk. First, it checks if the file
 returns a dataframe.
}
\examples{

}

