% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS Data File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string specifying the path to the FARS data file.}
}
\value{
A tibble containing the data from the specified file.
}
\description{
This function reads a Fatality Analysis Reporting System (FARS) dataset file.
The file must exist in the working directory.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
