% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{FARS Read Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of either intigers or strings correspondig to
years for which the data set is available.}
}
\value{
A list of FARS tibbles.
}
\description{
Create a list of year tibbles.
}
\details{
For each year passed into the function, create an tibble in a list
that contains just the years and months from the dataset for those
years.
}
\examples{
fars_read_years(c("2014", 2015))
}
