% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{extract the data for given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years to be filtered from the files.}
}
\value{
returns the dataframe includings the years given in the input.
}
\description{
This function get vector of years, load the data belonging to that date from
the file and returns a datarame of including the given years. It prints
invalid year if the year is not included in the data.
}
\examples{

}

