% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data for Multiple Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers representing the years of interest.}
}
\value{
A list of tibbles, each containing the month and year data for a specified year.
Invalid years will result in warnings.
}
\description{
This function reads FARS data for multiple years and extracts the month and year columns.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014))
}
}
