% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS Summarize Years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of either intigers or strings correspondig to
years for which the data set is available.}
}
\value{
A tibble of 12 rows (one per month) a column for month
    number, and as many year columns as were passed to the
    function.
}
\description{
Produce a summary of the number of accidents each month in the given years.
}
\details{
For each year passed into the function, create a tibble listing
    the number of accidents per month for that year.

Errors: An invalid year will produce an "invalid year" warning,
    and result in a empty column for that year.
}
