% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{returns the count of months and years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a character string or integer list for set of files beloging to
years given in the input.}
}
\value{
returns the count of data points for each month and year given in the
input in a wide format.
}
\description{
The function reads the list file given in parameter years, binds the rows of
each dataframe and returns the count of years and month.
}
\examples{

}

