% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS Data by Year and Month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers representing the years of interest.}
}
\value{
A tibble summarizing the number of observations by year and month.
}
\description{
This function summarizes the number of FARS observations by year and month.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014))
}
}
