% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{FARS make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or integer representation of a year.}
}
\value{
A string matching a valid file name for the dataset, e.g.,
    for 2015, \code{make_filename(2015)} will return
    "./data/accident_2015.csv.bz2"
}
\description{
From the year, creates a valid file name to work with FARS
dataset.
}
\examples{
make_filename(2014)
make_filename("2015")
}
