% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{print the filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{character string or number of year.}
}
\value{
returns the modified filename with the given year added to it.
}
\description{
The function get integer as input add return the modified filename.
}
\examples{

}


