% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a FARS File Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer representing the year.}
}
\value{
A string containing the FARS file name for the given year.
}
\description{
This function generates a standard FARS file name for a given year.
}
\examples{
make_filename(2013)
}
